zynqmp-production: Add support for the kv260 production SOM
Add basic support for kv-260 production SOM.

It's worth noting that the board has an eMMC and supports OP-TEE,
so we could use it to store the EFI variables.  However there's no
trustzone controller enabled (yet) and currently OP-TEE is unable to
program the RPMB key (which seems unique per device).  So store the
EFI variables in the SD card.

While at it move the common files currently named zynqmp-starter-*
to zynqmp-* in order to unify common files. Also bump U-Boot version
to pull in Xilinx support from -master and add the needed files to
compile op-tee in the future

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>