meta-ts: tfa/optee: stm32mp15: disable optee reserved shared memory
Disable OP-TEE reserved shared memory that is a specific reserved
non-secure memory used by REE and TEE but is used by Linux or U-Boot
for that platform since OP-TEE dynamic shared memory is enable.

CFG_CORE_RESERVED_SHM=n disables the memory in OP-TEE. This change adds
a patch for stm32mp1 optee-os.git to fix a related issue. the fix will
be merged in OP-TEE 3.19.0 [1] but is needed for 3.18.0.

STM32MP15_OPTEE_RSV_SHM=0 disables use OP-TEE reserved shared memory
but TF-A v2.7 does not implement that switch hence this change picks the
related commit mainline repository [2] into a local patch file in TF-A
recipe.

Link: [1] https://github.com/OP-TEE/optee_os/pull/5568
Link: [2] https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/commit/?id=722ca35ecc1c5de8682ca8df315a6369d0c21946
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>