u-boot-zynqmp-production.inc: append to SRC_URI and DEPENDS with +=
The :append can only be removed with :remove and is not nice
when several layers need to change a recipe. :append is mostly
used with machine, distro etc specific appends. For generic
appends += is enough and friendly for users who need to amend
the recipes and bbappends in other layers

Signed-off-by: Mikko Rapeli <mikko.rapeli@linaro.org>