meta-ts: stm32mp15: add partitions for u-boot and EFI variables
This change adds an ESP (EFI System Partition) to STM32MP157C-DK2
firmware boot image so that U-Boot/UEFI can store the EFI persistent
variables as required by ACS test suite.

This change also adds a storage area for U-Boot environment as it is
also quite convenient to be able to save U-Boot variables for test
and development purpose. This applies to bothj STM32MP157C-EV1 and -DK2.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>