rpi4-firmware: remove SRC_URI:append's
Just set the SRC_URI once. :append's are difficult to overwrite
with bbappends in other layers so if += works, then that should be
preferred. Note that this style of download URL from github.com
causes warnings from yocto since github is known to change the
content of tar balls which invalidates the checksums. This breaks
download caching.

Changing to use git with shallow clone option would be
the correct fix but download was so slow that I had to
abandone the build after 30 minutes. Maybe some other time...

Signed-off-by: Mikko Rapeli <mikko.rapeli@linaro.org>